move common_app
