+++ ydb/core/persqueue/pqrb/read_balancer_app.cpp
@@ -2,7 +2,7 @@
 
 #include "read_balancer__balancing.h"
 
-#include <ydb/core/persqueue/common_app.h>
+#include <ydb/core/persqueue/common/common_app.h>
 
 namespace NKikimr::NPQ {
