+++ ydb/core/persqueue/pqrb/ya.make
@@ -19,6 +19,7 @@ PEERDIR(
     ydb/core/base
     ydb/core/engine/minikql
     ydb/core/persqueue/events
+    ydb/core/persqueue/common
     ydb/core/persqueue/partition_key_range
     ydb/library/logger
 )
