+++ ydb/core/persqueue/pqtablet/partition/ya.make
@@ -22,6 +22,7 @@ PEERDIR(
     ydb/core/base
     ydb/core/engine/minikql
     ydb/core/persqueue/events
+    ydb/core/persqueue/common
     ydb/core/persqueue/partition_key_range
     ydb/core/persqueue/pqtablet/common
     ydb/library/logger
