+++ ydb/core/persqueue/pqtablet/pq_impl_app.cpp
@@ -2,7 +2,7 @@
 #include "pq_impl_types.h"
 
 #include <ydb/core/persqueue/pqtablet/common/logging.h>
-#include <ydb/core/persqueue/common_app.h>
+#include <ydb/core/persqueue/common/common_app.h>
 
 #include <ydb/library/protobuf_printer/security_printer.h>
